Refactor subcommands so they can be called from invoke
Rather than creating an instance variable for the args Namespace
object, pass the arguments around between methods as keyword
arguments.

The benefit of this is that subcommands can now be called by
directly invoking their method, if needed. This is helpful
is the caller isn't the main() method, but something else like
an Invoke task (akin to a target in a Makefile):
http://www.pyinvoke.org/