Use logging
Introduce the logging module, in order to provide a simple facility
for adding log messages. Also, add two global CLI options:

* --verbose or -v, which raises the log level from the default,
  WARNING, to INFO. If given twice, the log level is instead
  set to DEBUG.
* --quiet or -q, which lowers the log level from WARNING to ERROR.

The root logger uses sys.stderr as its output stream, which means that
olx subcommands can still send text that is meant to be piped or parsed
to stdout, and logging should not intervene with that.

Finally, also support OLX_LOG_LEVEL as an environment variable.
OLX_LOG_LEVEL can be any named log level understood by the logging
module; case is unimportant. This is primarily intended for use in
CI pipelines where you may want to default to debug logging for
build logs. To do so, set OLX_LOG_LEVEL=debug, and you won't need
to bother with the -vv flag.