Make new-run fail if base branch isn't clean
The fact that new-run does "git add ." is troublesome in that it may
cause content to be included in a course run that isn't meant to be
there.

Thus, if we are configured to create a new branch, check that the
base branch is clean (meaning: "git status --porcelain" returns
the empty string) before creating a new branch.