Add --wait option for upload
The course upload REST API immediately returns with HTTP 200 (OK)
after the course archive has been received in full, and the course
archive is then processed asynchronously in the background.

This is perfectly fine default behavior, but we might want to instead
wait until the course is actually available, meaning that its task
has progressed through the Unpacking, Updating, and Succeeded stages.

Thus, add a --wait option for olx upload, and fail in case of an
error in asynchronous task processing.