Call logging initialization from __init__
Previously, we would set up the logging module from the CLI's main()
method. However, this would mean that logging initialization would be
missed if we never called CLI.main(), but used an invoke task
instead. Thus, in that scenario the OLX_LOG_LEVEL environment variable
would not be honored.

Move the configuration of the logging module into the CLI class
__init__ method instead. That way, OLX_LOG_LEVEL=debug will enable
debug logging even when we're being called from an invoke task.