Make envar parsing more suitable for invoke
Previously, we would apply the OLX_* environment variables as defaults
to the CLI parser. That means that if we do not ever call the CLI, but
instead called a task from invoke, those values would never be
applied.

Move environment variable parsing to a location where it works both
from a CLI invocation, and from an invoke task.