Fix "fatal: Needed a single revision" when testing for branch
"fatal: Needed a single revision" is the error message that git
rev-parse gives us when testing for a branch and that branch does not
exist. It's not "fatal" for our purposes at all, and pretending
otherwise will confuse users, so suppress that error message by
invoking "git rev-parse -q" instead.